requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.4
disnake-py-interactions==4.1.0
disnake.py==1.7.3
emoji==1.7.0
frozenlist==1.3.0
idna==3.3
install==1.3.5
multidict==6.0.2
orjson==3.6.7
python-dotenv==0.19.2
typing_extensions==4.1.1
yarl==1.7.2